+++ Carbon.Registry/Functions/ConvertTo-CarbonPermissionsApplyTo.ps1
@@ -1,0 +1,34 @@
+
+function ConvertTo-CarbonPermissionsApplyTo
+{
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory, ValueFromPipeline)]
+        [AllowEmptyString()]
+        [AllowNull()]
+        [ValidateSet('KeyOnly', 'KeyAndSubkeys', 'SubkeysOnly')]
+        [String] $ApplyTo
+    )
+
+    begin
+    {
+        Set-StrictMode -Version 'Latest'
+        Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
+
+        $map = @{
+            'KeyOnly' = 'ContainerOnly';
+            'KeyAndSubkeys' = 'ContainerAndSubcontainers';
+            'SubkeysOnly' = 'SubcontainersOnly';
+        }
+    }
+
+    process
+    {
+        if (-not $ApplyTo)
+        {
+            return
+        }
+
+        return $map[$ApplyTo]
+    }
+}